put factorial return as long and create permutationlinkList

+++ src/main/java/org/shapleyvalue/core/ShapleyValue.java
@@ -39,7 +39,7 @@ public class ShapleyValue {
 		if(logger.isDebugEnabled()) logger.debug("ShapleyValue calculate started");
 		
 		int size = cfunction.getNbPlayers();
-		int factorielSize = FactorialUtil.factorial(size);
+		long factorielSize = FactorialUtil.factorial(size);
 		List<List<Integer>> permutations = Permutations.getAllPermutation(size);
 		
 		for(int i=1; i<=size; i++) {
