+++ src/main/java/org/shapleyvalue/util/FactorialUtil.java
@@ -2,7 +2,7 @@ package org.shapleyvalue.util;
 
 public class FactorialUtil {
 
-	public static int factorial(int n) {
+	public static long factorial(long n) {
 		return n > 1 ? n * factorial(n - 1) : 1;
 	}
